+++ packages/log/src/log/modules/MessageConstructor/traps/deleteProperty.ts
@@ -1,4 +1,4 @@
-import { Design } from "../../../types";
+import { Design, Message, ProxyTrap } from "../../../types";
 import { Configuration } from "../../Configuration";
 
 export default function (
@@ -8,13 +8,25 @@ export default function (
     target: object,
     property: PropertyKey
 ): string[] {
-    const SD = Configuration
-        .getPreferredOptions(target, property)[Configuration.constants.STRING_DECORATOR];
-    return [
+    const options = Configuration.getPreferredOptions(target, property);
+    const SD = options[Configuration.constants.STRING_DECORATOR];
+
+    const msg: Message = [
         SD.getAssembledField(
             `Access delete ${SD.fieldLabel("operator")}`,
             `delete ${SD.objectLike(target, true)}[${SD.string(property as string)}];`
         ),
         SD.getAssembledField("Delete result", SD.boolean(result)),
     ];
+
+    msg.logData = {
+        proxyTrap: ProxyTrap.DELETE_PROPERTY,
+        propertyKey: property,
+        target,
+        trapResult: result,
+        design,
+        options,
+    };
+
+    return msg;
 }