+++ packages/log/src/log/modules/MessageConstructor/traps/ownKeys.ts
@@ -1,4 +1,4 @@
-import { Design, Message } from "../../../types";
+import { Design, Message, ProxyTrap } from "../../../types";
 import { Configuration } from "../../Configuration";
 
 export default function (
@@ -7,13 +7,13 @@ export default function (
     innerArgs: any[],
     target: object
 ): string[] {
-    const SD = target[Configuration.constants.OPTIONS][Configuration.constants.STRING_DECORATOR];
-    const msg: Message = [];
+    const options = Configuration.getPreferredOptions(target);
+    const SD = options[Configuration.constants.STRING_DECORATOR];
 
-    msg.push(
+    const msg: Message = [
         SD.getAssembledField("Get own keys of", SD.getLoggerNameString(target)),
-        SD.getAssembledField("Result", SD.logArrayProperties(result))
-    );
+        SD.getAssembledField("Result", SD.logArrayProperties(result)),
+    ];
 
     msg.extensibleObjects = [
         {
@@ -21,6 +21,14 @@ export default function (
             value: result,
         },
     ];
+
+    msg.logData = {
+        proxyTrap: ProxyTrap.OWN_KEYS,
+        target,
+        trapResult: result,
+        design,
+        options,
+    };
 
     return msg;
 }