+++ packages/log/src/log/modules/MessageConstructor/traps/set.ts
@@ -1,6 +1,6 @@
 import { isFunction, isSymbol } from "@js-utilities/typecheck";
 
-import { Design, DesignType, Message } from "../../../types";
+import { Design, DesignType, Message, ProxyTrap } from "../../../types";
 import { resolveDesignType } from "../../../utils";
 import { Configuration } from "../../Configuration";
 
@@ -60,6 +60,16 @@ export default function (
             name: "Set value",
         },
     ];
+
+    msg.logData = {
+        proxyTrap: ProxyTrap.SET,
+        propertyKey: property,
+        target,
+        trapResult: result,
+        design,
+        valueToSet: value,
+        options: preferredOptions,
+    };
 
     return msg;
 }