+++ packages/log/src/log/modules/MessageConstructor/traps/setPrototypeOf.ts
@@ -1,4 +1,4 @@
-import { Design, Message } from "../../../types";
+import { Design, Message, ProxyTrap } from "../../../types";
 import { Configuration } from "../../Configuration";
 
 export default function (
@@ -8,17 +8,17 @@ export default function (
     target: object,
     prototype: any
 ): string[] {
-    const SD = target[Configuration.constants.OPTIONS][Configuration.constants.STRING_DECORATOR];
+    const options = Configuration.getPreferredOptions(target);
+    const SD = options[Configuration.constants.STRING_DECORATOR];
     const res = Boolean(result) ? "succeed" : "failed";
-    const msg: Message = [];
 
-    msg.push(
+    const msg: Message = [
         SD.getAssembledField(
             `Call ${SD.methodName("setPrototypeOf")} ${SD.fieldLabel("on")}`,
             SD.getLoggerNameString(target)
         ),
-        SD.getAssembledField(`Set ${res}`, SD.unknownType(prototype, true))
-    );
+        SD.getAssembledField(`Set ${res}`, SD.unknownType(prototype, true)),
+    ];
 
     msg.extensibleObjects = [
         {
@@ -26,6 +26,15 @@ export default function (
             value: prototype,
         },
     ];
+
+    msg.logData = {
+        proxyTrap: ProxyTrap.SET,
+        target,
+        trapResult: result,
+        design,
+        options,
+        valueToSet: prototype,
+    };
 
     return msg;
 }