+++ packages/log/src/log/types/Configuration.ts
@@ -3,6 +3,7 @@ import { DecoratorType } from "./Decorator";
 import { Design } from "./Design";
 import { Configuration } from "../modules/Configuration";
 import { StringDecorator } from "../modules/StringDecorator";
+import { LogData } from "./Proxy";
 
 export type LoggerOptions = {
 
@@ -20,14 +21,6 @@ export type LoggerOptions = {
      * @default true
      */
     logName?: boolean;
-
-    /**
-     * If provided, will be executed on each log.
-     * If condition isn't met, log won't be executed.
-     *
-     * @default undefined
-     */
-    condition?: (propKey: PropertyKey) => boolean;
 
     /**
      * Console method which will be used to output messages.
@@ -57,6 +50,16 @@ export type LoggerOptions = {
      * @default { logTimeStamp: true, logSuffix: true }
      */
     loggerOptions?: InstanceMessageLoggerOptions;
+
+    /**
+     * Will be invoked on every log, right before writing to console.
+     * Can prevent log by returning "false".
+     *
+     * @param {LogData} logData   collected data of proxy trap
+     *
+     * @return {Boolean} log message or no
+     */
+    logInterceptor?(logData: LogData): boolean;
 
     /**
      * Flag for enabling/disabling logging.
@@ -299,6 +302,7 @@ export type LoggerTheme = {
 
 export type Message = string[] & {
     extensibleObjects?: ExtensibleObject[];
+    logData?: LogData;
 };
 
 export interface MessageHandler {