+++ packages/log/src/log/utils/other.ts
@@ -5,10 +5,9 @@ import { isBrowser } from "./environment";
 // we will create require fn dynamically to avoid
 // resolving "require()" calls by bundlers (webpack etc.)
 // we will need this only in node env, so simply mock it with dummy function in browsers
-// tslint:disable-next-line:no-eval
 export const requireFunc = isBrowser()
     ? () => void 0
-    : eval("require");
+    : eval("require"); // tslint:disable-line:no-eval
 
 export const now = () => (
     isBrowser()