+++ packages/log/CHANGELOG.md
@@ -1,6 +1,24 @@
 # 0.1.4 (not released yet)
 
 * highlight `bound` keyword for bound functions
+* add `logInterceptor` callback, which will be invoked before every log. Also, returned value from this callback will decide whether to log message to console or not. `logInterceptor` accepts an object with a lot of data of proxied call:
+
+```typescript
+type LogData = {
+    proxyTrap: ProxyTrap;
+    propertyKey?: PropertyKey;
+    trapResult?: any;
+    target?: any;
+    options?: LoggerOptions;
+    message?: string;
+    decoratedMessage?: string;
+    valueToSet?: any;
+    descriptor?: PropertyDescriptor;
+    thisArg?: any;
+    args?: any[];
+    design?: Design;
+};
+```
 
 # 0.1.3